fix: 修复全部已识别问题 + 补齐构建体系 (可编译工程)
编译/链接级:
- freertos.c: 修复 vControlTask 前注释块丢失定界符的语法错误;
  g_att/g_att_ready 先用后声明
- i2c.c/mpu6050_i2c.c: 消除 hi2c1 重复定义, 引脚统一为 PB8/PB9(上拉)
- stm32f4xx_it.c: SVC/PendSV 交由 FreeRTOS 端口提供;
  SysTick_Handler 双跳 (HAL_IncTick + xPortSysTickHandler) 解决
  RTOS 接管后 HAL 时基冻结
- FreeRTOSConfig.h: 实现 Malloc/StackOverflow Hook, 关闭 TickHook,
  修正定时器任务优先级注释

功能级:
- main.c: 补全初始化链 (GPIO→电机PWM→UART→I2C→MPU重试→IWDG→RTOS),
  原空 while(1) 什么都不做
- usart.c: 补 printf 重定向 (_write) + USART1 中断接收→字节队列
- freertos.c: 新增 Comm 任务打通 BLE 指令链 (SETPOINT/ARM/EMERGENCY/PING);
  零偏校准移入 ImuRead 任务消除 I2C 并发竞争; 姿态任务 200→500Hz
  匹配采样率 (原队列恒满丢帧); 新增姿态丢失 0.5s failsafe; IWDG 喂狗
- ble_protocol.h: 新增 BLE_CMD_ARM (0x24)
- control.c: 新增 Control_KeepAlive 心跳

数学/逻辑错误:
- data_process.c: 修复垂直加速度用了转置旋转矩阵 (两交叉项符号反);
  Mahony/滤波链统一 500Hz (dt=2ms); 校准只用有效样本求均值
- motor_virtual.c: 修复 roll 力矩用了对角差 (应为左右差) 及 yaw 符号
- mpu6050_i2c.h: 修复 MPU_ADDR 宏优先级陷阱; 单位注释 rad/s
- safety.c: printf 格式 %lu 显式转换

构建体系 (新增):
- Makefile (arm-none-eabi-gcc, nano + _printf_float, gc-sections)
- STM32F411CEUX_FLASH.ld 链接脚本
- vendor: ST HAL 子集 / CMSIS 5 / F4 设备头+启动文件 / FreeRTOS 内核
  (tasks,queue,list,timers + CM4F 端口 + heap_4)
- Tools/host_test: 主机端控制逻辑回归 (27 断言)
- README.md + .gitignore

验证: 全树 39 源文件 -fsyntax-only 通过 (应用+HAL+FreeRTOS 真实头文件);
Makefile 干跑正常; 控制逻辑回归 27/27 通过

+++ Project/Drone_F411/Core/Src/safety.c
@@ -68,7 +68,7 @@ void Safety_Update(float roll, float pitch, float yaw,
     if (g_safety.armed && now_ms - g_safety.last_cmd_ms > g_safety.cmd_timeout_ms) {
         g_safety.status = SAFE_LOST_CONTROL;
         printf("[SAFETY] LOST CONTROL: %lums no cmd\r\n",
-               now_ms - g_safety.last_cmd_ms);
+               (unsigned long)(now_ms - g_safety.last_cmd_ms));
         return;
     }
 
@@ -119,7 +119,7 @@ void IWDG_Init(uint32_t reload_ms) {
     IWDG->RLR = reload_ms;       /* 重载值 */
     IWDG->KR  = 0xCCCC;          /* 启动 */
     IWDG->KR  = 0xAAAA;          /* 首次喂狗 */
-    printf("[SAFETY] IWDG started: %lums timeout\r\n", reload_ms);
+    printf("[SAFETY] IWDG started: %lums timeout\r\n", (unsigned long)reload_ms);
 }
 
 void IWDG_Feed(void) {
